backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.26.2
opencv-python-headless==4.8.1.78
matplotlib==3.8.2
scikit-learn==1.3.2
gunicorn==21.2.0
